crud operations.

+++ gauresh_BackEnd/src/main/java/com/niit/gauresh_backend/model/Product.java
@@ -73,7 +73,7 @@ public class Product implements Serializable {
 	public void setDescription(String description) {
 		this.description = description;
 	}
-	public double getUnitPrice() {
+	public int getUnitPrice() {
 		return unitPrice;
 	}
 	public void setUnitPrice(int unitPrice) {
